move create-config to separate file
